refactor(chat): extract chat_manager/history.rs (#292)
Move the history, context, and degraded-recovery delegation methods
off the facade impl into their own submodule.
