+++ packages/core-rs/crates/mainframe-chat/src/chat_manager.rs
@@ -66,6 +66,7 @@ mod deps_permission;
 mod deps_recovery;
 mod errors;
 mod external_facade;
+mod history;
 mod lifecycle_api;
 mod reads;
 mod send;
@@ -112,126 +113,6 @@ pub struct ChatManager {
 }
 
 impl ChatManager {
-    // ── history + context reads ──────────────────────────────────────────────
-
-    /// Cached messages, falling back to a one-shot on-disk history load (Claude
-    /// `--resume` JSONL). The load remaps the embedded Claude sessionId back to the
-    /// Mainframe chatId and restores any pending permission from history.
-    pub async fn get_messages(&self, chat_id: &str) -> Vec<ChatMessage> {
-        self.lifecycle.await_loading(chat_id).await;
-
-        let cached = self
-            .messages
-            .lock()
-            .unwrap_or_else(|e| e.into_inner())
-            .get(chat_id)
-            .cloned();
-        if let Some(cached) = cached
-            && !cached.is_empty()
-        {
-            return cached;
-        }
-
-        let Some(session) = self.history_session(chat_id) else {
-            return Vec::new();
-        };
-        match session.load_history().await {
-            Ok(history) => {
-                let remapped = remap_history(history, chat_id);
-                if !remapped.is_empty() {
-                    self.messages
-                        .lock()
-                        .unwrap_or_else(|e| e.into_inner())
-                        .set(chat_id, remapped.clone());
-                    self.permissions
-                        .lock()
-                        .unwrap_or_else(|e| e.into_inner())
-                        .restore_pending_permission(chat_id, &remapped);
-                }
-                remapped
-            }
-            Err(_) => Vec::new(),
-        }
-    }
-
-    /// Load messages from disk, bypassing the in-memory cache (session-files route
-    /// needs subagent file changes absent from the cache during an active session).
-    pub async fn get_messages_from_disk(&self, chat_id: &str) -> Vec<ChatMessage> {
-        let Some(session) = self.history_session(chat_id) else {
-            return Vec::new();
-        };
-        match session.load_history().await {
-            Ok(history) => remap_history(history, chat_id),
-            Err(err) => {
-                tracing::warn!(?err, chat_id, "getMessagesFromDisk failed");
-                Vec::new()
-            }
-        }
-    }
-
-    /// Display history + transcript presence in one typed result, so the REST
-    /// route (and the UI) can tell an empty thread from a deleted transcript.
-    /// Reconciling here persists flag flips and broadcasts `chat.updated`.
-    pub async fn get_display_messages(&self, chat_id: &str) -> ChatHistoryPayload {
-        let raw = self.get_messages(chat_id).await;
-        let categories = self.deps.get_tool_categories(chat_id);
-        let messages = self
-            .deps
-            .prepare_messages_for_client(&raw, categories.as_ref());
-        let transcript_missing = match self.get_chat(chat_id) {
-            Some(mut chat) => self.reconcile_transcript(&mut chat).await,
-            None => false,
-        };
-        ChatHistoryPayload {
-            messages,
-            transcript_missing,
-            workflow_runs: Vec::new(),
-        }
-    }
-
-    /// Reconcile the persisted `transcriptMissing` flag against the transcript file
-    /// on disk.
-    pub async fn reconcile_transcript(&self, chat: &mut Chat) -> bool {
-        let wrapper = self.recovery_wrapper();
-        crate::transcript_presence::reconcile_transcript_presence(&wrapper, chat).await
-    }
-
-    /// Forget the dead CLI session so the next send spawns fresh in the same chat row.
-    pub async fn continue_here(&self, chat_id: &str) -> Result<(), DegradedRecoveryError> {
-        let wrapper = self.recovery_wrapper();
-        crate::degraded_recovery::continue_here(&wrapper, chat_id).await
-    }
-
-    /// Detach the chat from its deleted worktree and rebind it to the project root.
-    pub async fn continue_in_project_root(
-        &self,
-        chat_id: &str,
-    ) -> Result<(), DegradedRecoveryError> {
-        let wrapper = self.recovery_wrapper();
-        crate::degraded_recovery::continue_in_project_root(&wrapper, chat_id).await
-    }
-
-    /// Re-add the deleted worktree at its stored path from the stored branch (409 when branch gone).
-    pub async fn recreate_worktree(&self, chat_id: &str) -> Result<(), DegradedRecoveryError> {
-        let wrapper = self.recovery_wrapper();
-        crate::degraded_recovery::recreate_chat_worktree(&wrapper, chat_id).await
-    }
-
-    /// Build a stateless history-load session for `chat_id`, or `None` when the chat
-    /// has no Claude session / adapter / project. Mirrors `getMessages`'s guard chain.
-    fn history_session(&self, chat_id: &str) -> Option<Arc<dyn AdapterSession>> {
-        let chat = self.get_chat(chat_id)?;
-        build_history_session(&self.deps, &chat, chat_id)
-    }
-
-    pub async fn get_session_context(&self, chat_id: &str, project_path: &str) -> SessionContext {
-        let session = self.get_session_for_chat(chat_id);
-        let adapter_id = self.get_chat(chat_id).map(|c| c.adapter_id);
-        self.deps
-            .get_session_context(chat_id, project_path, session, adapter_id)
-            .await
-    }
-
     // ── config + worktree delegations ────────────────────────────────────────
 
     pub async fn update_chat_config(
